This commit includes a number of significant changes:
(1) Each stage is wrapped in an outer loop. This continues until no move can be made for a single batch. Moves are then tested to ensure they improve fairness. So in fact our loop continues until no fairness improvement can be made. This is important because it allows the algorithm to terminate should it not be able to complete (certain configurations cannot create fairness, due to the rack and partition constraints). By choosing a batch to be all current above par replicas we ensure a sensible termination condition.

(2) There are now two definitions of fairness: ceil() & floor() of the fairness value. BelowPar is always defined with ceil(). AbovePar is always defined with floor(). Thus, where fairness is a fractional value, there will be an overlap between abovePar and belowPar. This is required as the algorithm is not symmetric (it starts at with abovePar brokers then searches belowPar brokers). Consider the an arrangmeent where the replica counts on each broker are (1, 1, 2, 3, 3). There are 11 replicas over 5 brokers so fairness = 2.2. Thus ceil(fairvalue) -> 3 and would not achieve fairness. Rerunning in the opposite direction (belowPar->abovePar) also does not help  as there are no above par options to move to. Hence we choose to apply ceil/floor to the abovePar/belowPar.

(3) Fixed bug with rackConstraint when working in “inside rack” (aka BrokerView) mode. Partitions needed to be filtered by rack but weren’t being.